#68 Move codes to be deprecated
